fix(limelight): accept float ts field from /results
Limelight returns ts as a JSON float with sub-millisecond precision
(e.g. "ts":588916.439519), not an integer. The previous *int64
decode target rejected every poll with "cannot unmarshal number X.Y
into Go struct field LimelightResponse.ts of type int64", which
manifested as reachable=false for the entire match.

Decode into *float64 (accepts both int and float JSON numbers) and
emit uptime_ms as a "double" TopicValue instead of "int". WPILog
export already handles the "double" type correctly via
EncodeValue's explicit double case, so AdvantageScope still plots
the topic cleanly.

Test pins the exact number from the wild-failure log so this can't
regress.

+++ internal/limelight/monitor.go
@@ -147,11 +147,14 @@ func (m *Monitor) tick(ctx context.Context) {
 }
 
 // limelightResponse is the minimal shape we care about from the
-// /results JSON body. A pointer distinguishes "field absent" from
-// "field present with value 0" (a freshly-booted Limelight is at
-// ts=0 and is a legitimate datapoint, not an error).
+// /results JSON body. The ts field is "milliseconds since boot" but
+// Limelight returns it as a float with sub-millisecond precision
+// (e.g. 588916.439519), so we decode into *float64. A pointer
+// distinguishes "field absent" from "field present with value 0"
+// (a freshly-booted Limelight is at ts=0 and is a legitimate
+// datapoint, not an error).
 type limelightResponse struct {
-	TS *int64 `json:"ts"`
+	TS *float64 `json:"ts"`
 }
 
 // poll issues one request to octet's Limelight. On success it sends
@@ -192,7 +195,7 @@ func (m *Monitor) poll(ctx context.Context, octet int) {
 		return
 	}
 
-	m.send(uptimeTopic(octet), "int", *body.TS)
+	m.send(uptimeTopic(octet), "double", *body.TS)
 	m.recordResult(octet, url, true, "")
 }
